Refactored rooms routes for unit testing.
